chore(nhost_flutter_graphql): Extract provider code to src
I'm about to add a second file, so introducing the private namespace
seemed valuable.
